appstructure is fixed, css code is converted to scss, and other stuff is fixed
