--- NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/CSS/App.css
@@ -1,9 +1,0 @@
-a {
-  text-decoration: none;
-  line-height: 2em;
-  font-weight: 500;
-}
-ul {
-  list-style-type: none;
-  padding: 0;
-}
