+++ NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/CSS/Seeds.css
@@ -1,4 +1,4 @@
-.payWithSeeds{
+.popup-container {
   width: 100vw;
   height: 100vh;
   display: grid;
@@ -9,11 +9,11 @@
   z-index: 10;
 	background-color: hsla(240, 8%, 12%, 0.7);
 }
-.payWithSeeds h1{
+.popup-container h1{
   color: #fff;
   font-weight: 400;
 }
-.payWithSeeds-select{
+.popup-container-select{
   font-size: 1.5rem;
   color: #fff;
   background-color: transparent;
@@ -23,37 +23,37 @@
   border: 1px solid #909092;
   width: auto;
 }
-.payWithSeeds input {
+.popup-container input {
   border: 1px solid #909092;
   font-size: 1.15rem;
   padding: 0.25rem 1rem;
   border-radius: 5px;
   color: #909092;
 }
-.payWithSeeds label{
+.popup-container label{
   color: #fff;
   padding: 0.25rem 1rem;
 }
-.payWithSeeds-popup{
+.popup-container-popup{
   background-color: hsla(240, 8%, 12%, 0.8);
   box-shadow: 0 0 2.5rem rgba(136, 133, 133, 0.308);
   padding: 2rem;
   border-radius: 1rem;
   display: grid;
 }
-.payWithSeeds-fields{
+.popup-container-fields{
   display: grid;
   row-gap: 1rem;
   grid-template-columns: auto 1fr;
 }
-.payWithSeeds-button-align{
+.popup-container-button-align{
   display: grid;
   margin-top: 2rem;
   grid-template-columns: auto auto;
   justify-content: right;
   column-gap: 1rem;
 }
-.payWithSeeds-row{
+.popup-container-row{
   display: grid;
   grid-template-columns: repeat(2, auto 1fr);
   margin-top: 1rem;
@@ -64,28 +64,28 @@
   margin-top: 1rem;
 }
 @media (max-width: 800px) {
-  .payWithSeeds{
+  .popup-container{
     justify-content: initial;
   }
-  .payWithSeeds-label{
+  .popup-container-label{
     text-align: left;
   }
-  .payWithSeeds-popup{
+  .popup-container-popup{
     grid-template-columns: 80%;
     justify-content: center;
   }
-  .payWithSeeds-fields, .payWithSeeds-row, .sendInvite-row{
+  .popup-container-fields, .popup-container-row, .sendInvite-row{
     grid-template-columns: 100%;
     row-gap: 0.5rem;
   }
-  .payWithSeeds-button-align{
+  .popup-container-button-align{
     grid-template-columns: 1fr 1fr;
     justify-content: initial;
   }
 }
 
 @media (max-width: 550px) {
-  .payWithSeeds-popup{
+  .popup-container-popup{
     grid-template-columns: 100%;
   }
 }
