+++ NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/Components/App.js
@@ -1,47 +1,147 @@
-import Navbar from './NavBar/Navbar'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import PayWithSeeds from './seeds/PayWithSeeds'
-import SendInvite from './seeds/SendInvite'
+import React from "react";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
-// ============= All of the css imports ===========
-import '../CSS/App.css'
-import '../CSS/Navbar.css'
-import '../CSS/Login.css'
-import '../CSS/SideNav.css'
-import '../CSS/Alert.css'
-import '../CSS/Seeds.css'
-// ================================================
-
-import Karma from "../pages/Karma"
-import Home from "../pages/Home"
+import SideNav from "./common/SideNav";
+import Navbar from "./common/Navbar";
+import PayWithSeeds from "./pages/seeds/PayWithSeeds";
+import SendInvite from "./pages/seeds/SendInvite";
+import Karma from "./pages/karma/Karma";
+import Home from "../pages/Home";
+import Login from "./Login";
+import Signup from "./Signup";
 
 
-function App() {
-  return (
-    <Router>
-      <header className="App-header">
-        <Navbar />
-      </header>
-      <Switch>
-        <Route path="/payWithSeeds">
-          <PayWithSeeds />
-        </Route>
-        <Route path="/donateWithSeeds">
-          <PayWithSeeds seedType="Donate" />
-        </Route>
-        <Route path="/rewardWithSeeds">
-          <PayWithSeeds seedType="Reward" />
-        </Route>
-        <Route exact path="/">
-          <Home />
-        </Route>
-        <Route path="/sendInvite">
-          <SendInvite />
-        </Route>
-        <Route exact path="/karma" component={Karma} />
-      </Switch>
-    </Router>
-  )
+import "../assets/scss/general.scss";
+import "../assets/scss/style.scss";
+import "../css/Seeds.css";
+
+import axios from "axios";
+
+class App extends React.Component {
+    state = {
+        showSidebar: false,
+        showLogin: false,
+        showSignup: false,
+        user: null
+    }
+
+    componentDidMount() {
+        localStorage.getItem('user')
+
+        if (localStorage.getItem('user')) {
+            this.setState({ user: JSON.parse(localStorage.getItem('user')) })
+        }
+    }
+
+    setUserData = (data) => {
+        console.log(data);
+        this.setState({
+            user: data
+        })
+    }
+
+    toggleSidebar = () => {
+        this.setState({
+            showSidebar: !this.state.showSidebar
+        })
+    }
+
+    hidePopups = () => {
+        this.setState({
+            showLogin: false,
+            showSignup: false
+        })
+    }
+
+    hideLogin = () => {
+        this.setState({
+            showLogin: false,
+        })
+    }
+
+    hideSignup = () => {
+        this.setState({
+            showSignup: false,
+        })
+    }
+
+    showLogin = () => {
+        this.setState({
+            showLogin: true,
+            showSignup: false,
+        })
+    }
+
+    showSignup = () => {
+        this.setState({
+            showSignup: true,
+            showLogin: false
+        })
+    }
+
+    handleLogout = () => {
+        axios.post('https://api.oasisplatform.world/api/avatar/revoke-token', {
+            token: this.state.user.jwtToken
+        }).then(res => {
+            this.setState({ user: null })
+            localStorage.removeItem('user')
+            localStorage.removeItem('credentials')
+        }).catch(err => {
+            this.setState({ user: null })
+            localStorage.removeItem('user')
+            localStorage.removeItem('credentials')
+        })
+    }
+
+    render() { 
+        return (
+            <div className="main-container">
+                <Router>
+                    <header>
+                        <Navbar 
+                            showSidebar={this.state.showSidebar} 
+                            toggleSidebar={this.toggleSidebar}
+                            showLogin={this.showLogin} 
+                            showSignup={this.showSignup} 
+                            handleLogout={this.handleLogout}
+                            user={this.state.user}
+                        />
+                        <SideNav showSidebar={this.state.showSidebar} toggleSidebar={this.toggleSidebar}  />
+                    </header>
+
+                    <div className="content-container">
+                        <Switch>
+                            <Route exact path="/home" component={Home} />
+                            <Route path="/pay-with-seeds" component={PayWithSeeds} />
+                            <Route path="/donateWithSeeds">
+                                <PayWithSeeds seedType="Donate" />
+                            </Route>
+                            <Route path="/rewardWithSeeds">
+                                <PayWithSeeds seedType="Reward" />
+                            </Route>
+                            <Route path="/send-invite" component={SendInvite} />
+                            <Route exact path="/karma" component={Karma} />
+                        </Switch>
+                    </div>
+                </Router>
+
+                <Login 
+                    className="custom-form"
+                    show={this.state.showLogin} 
+                    hide={this.hideLogin}
+                    change={this.showSignup} 
+                    setUserStateData={this.setUserData}
+                /> 
+
+                <Signup
+                    className="custom-form" 
+                    show={this.state.showSignup} 
+                    hide={this.hideSignup}
+                    change={this.showLogin} 
+                /> 
+            </div>
+        );
+    }
 }
 
 export default App;
