--- NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/Components/NavBar/Alert.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Alert = ({ type, message }) => (
-    <div className={`alert alert--${type}`}>
-        <p>{message}</p>
-    </div>
-);
-
-export default Alert;
