--- NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/Components/NavBar/SideNavData.js
@@ -1,219 +1,0 @@
-const SideNavData = [
-  {
-    title: "the oasis",
-    subNav: [
-      {
-        title: "about",
-        path: ""
-      },
-      {
-        title: "documentation",
-        path: ""
-      },
-      {
-        title: "our world",
-        path: ""
-      },
-    ]
-  },
-  {
-    title: "avatar",
-    subNav: [
-      {
-        title: "view avatar",
-        path: ""
-      },
-      {
-        title: "edit avatar",
-        path: ""
-      },
-      {
-        title: "search avatars",
-        path: ""
-      }
-    ]
-  },
-  {
-    title: "karma",
-    subNav: [
-      {
-        title: "View Curent Karma Weightings",
-        path: ""
-      }, {
-        title: "Vote For Karma Weightings",
-        path: ""
-      }, {
-        title: "View Avatar Karma",
-        path: ""
-      }, {
-        title: "View/Search Karma Akashic Records",
-        path: ""
-      }
-    ]
-  },
-  {
-    title: "data",
-    subNav: [
-      {
-        title: "Load Data",
-        path: ""
-      }, {
-        title: "Send Data",
-        path: ""
-      }, {
-        title: "Manage Data",
-        path: ""
-      }, {
-        title: "Cross-Chain Management",
-        path: ""
-      }, {
-        title: "Off-Chain Management",
-        path: ""
-      }, {
-        title: "Search Data",
-        path: ""
-      },
-    ]
-  },
-  {
-    title: "seeds",
-    subNav: [
-      {
-        title: "Pay With SEEDS",
-        path: "/payWithSeeds"
-      }, {
-        title: "Donate SEEDS",
-        path: "/donateWithSeeds"
-      }, {
-        title: "Reward SEEDS",
-        path: "/rewardWithSeeds"
-      }, {
-        title: "Invite To Join SEEDS",
-        path: "/sendInvite"
-      }, {
-        title: "Accept Invit to join seeds",
-        path: ""
-      }, {
-        title: "View SEEDS",
-        path: ""
-      }, {
-        title: "View Organisations",
-        path: ""
-      }, {
-        title: "Manage SEEDS",
-        path: ""
-      }, {
-        title: "Search Seeds",
-        path: ""
-      }
-    ]
-  },
-  {
-    title: "provider",
-    subNav: [
-      {
-        title: "View Providers",
-        path: ""
-      }, {
-        title: "Manage Providers",
-        path: ""
-      }, {
-        title: "Manage Auto-Replication",
-        path: ""
-      }, {
-        title: "Manage Auto-Fail-Over",
-        path: ""
-      }, {
-        title: "Manage Load Balancing",
-        path: ""
-      }, {
-        title: "View Provider Stats",
-        path: ""
-      }, {
-        title: "Compare Provider Speeds",
-        path: ""
-      }, {
-        title: "Search Providers",
-        path: ""
-      }, {
-        title: "Holochain",
-        path: ""
-      }, {
-        title: "SEEDS",
-        path: ""
-      }, {
-        title: "EOSIO",
-        path: ""
-      }, {
-        title: "Ethereum",
-        path: ""
-      }, {
-        title: "IPFS",
-        path: ""
-      }, {
-        title: "ThreeFold",
-        path: ""
-      }, {
-        title: "SOLID",
-        path: ""
-      }, {
-        title: "Activity Pub",
-        path: ""
-      }, {
-        title: "Mongo DB",
-        path: ""
-      }, {
-        title: "SQLLite",
-        path: ""
-      }, {
-        title: "Neo4j",
-        path: ""
-      }
-    ]
-  },
-  {
-    title: "nft",
-    subNav: []
-  },
-  {
-    title: "map",
-    subNav: []
-  },
-  {
-    title: "oapp",
-    subNav: []
-  },
-  {
-    title: "quest",
-    subNav: []
-  },
-  {
-    title: "mission",
-    subNav: []
-  },
-  {
-    title: "egg",
-    subNav: []
-  },
-  {
-    title: "game",
-    subNav: [
-      {
-        title: "View StarCraft 2 Leagues",
-        path: ""
-      }, {
-        title: "View StarCraft 2 Tournaments",
-        path: ""
-      }, {
-        title: "View StarCraft 2 Achievements",
-        path: ""
-      }
-    ]
-  },
-  {
-    title: "developer",
-    subNav: []
-  }
-]
-
-export default SideNavData;
