+++ NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/assets/images/loggedin.png
