+++ NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/assets/scss/general.scss
@@ -1,0 +1,232 @@
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+    font-family: sans-serif;
+}
+
+body {
+    margin: 0;
+    font-family: 'Titillium Web', sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    background: url("../images/bg.gif");
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-attachment: fixed;
+    background-position: top;
+}
+
+a {
+    text-decoration: none;
+}
+ul {
+    padding: 0;
+    li {
+        list-style: none;
+    }
+}
+  
+code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+}
+
+input[type=number]::-webkit-inner-spin-button, 
+input[type=number]::-webkit-outer-spin-button { 
+    -webkit-appearance: none;
+    -moz-appearance: none;
+    appearance: none;
+    margin: 0; 
+}
+input[type=number] {
+    -moz-appearance:textfield;
+}
+
+button {
+    text-decoration: none;
+    border: none;
+    background-color: transparent;
+}
+input {
+    background-color: transparent;
+}
+
+.link {
+    color: #1885c9;
+    &:hover {
+        color: #fff;
+        cursor: pointer;
+        transition: all 0.2s ease;
+    }
+}
+.link-inverse { 
+    color: #fff;
+    &:hover {
+        color: #1885c9;
+        cursor: pointer;
+        transition: all 0.2s ease;
+    }
+}
+.button {
+    color: #fff;
+    background-color: #1885c9;
+    padding: 0.5rem 1rem;
+    border-radius: 5rem;
+    cursor: pointer;
+    font-size: 1.15rem;
+    transition: all 200ms ease-in;
+    &:hover {
+        background-color: #52b2ee;
+    }
+}
+.button-inverse{
+    color: #1885c9;
+    padding: 0.5rem 1rem;
+    border-radius: 5rem;
+    cursor: pointer;
+    font-size: 1.15rem;
+    border: 1px solid #1885c9;
+    transition: all 200ms ease-in;
+    &:hover {
+        background-color: #52b2ee;
+        color: #fff;
+    }
+}
+
+.custom-modal {
+    span.form-cross-icon {
+        color: #fff;
+        position: absolute;
+        right: 20px;
+        top: 10px;
+        font-size: 24px;
+        font-weight: 300;
+        z-index: 10;
+        cursor: pointer;
+    }
+    .modal-content {
+        background-color: initial;
+        .modal-body {
+            padding: 0;
+        }
+    }
+}
+
+.text-link {
+    color: #1885c9;
+    cursor: pointer;
+}
+
+.custom-form {
+	background-color: hsla(240, 8%, 12%, 0.8);
+	border-radius: 8px;
+	padding: 2rem 3rem;
+    text-align: center;
+    position: relative;
+    p, h2 {
+        color: #fff;
+    }
+    h2 {
+        font-size: 32px;
+        margin: 0 0 15px;
+    }
+    .form-header {
+        margin: 0 0 35px;
+        p {
+            margin: 0;
+        }
+    }
+    .form-inputs {
+        .remember-me {
+            text-align: left;
+            color: #fff;
+            margin: 0 0 25px;
+            label {
+                cursor: pointer;
+                margin: 0;
+                input {
+                    margin-right: 10px;
+                }
+            }
+        }
+        .forgot-password {
+            text-align: left;
+            label {
+                margin: 0;
+                cursor: pointer;
+            }
+        }
+        .single-form-field {
+            display: grid;
+            margin: 0 0 25px;
+            label {
+                margin: 0 0 5px;
+                color: #fff;
+                display: block;
+                text-align: left;
+            }
+            input {
+                padding: 0 15px;
+                height: 40px;
+                border-radius: 4px;
+                margin: 0;
+                outline: none;
+                border: none;
+                background-color: #fff;
+            }
+            .have-icon {
+                position: relative;
+                display: grid;
+                input {
+                    padding-right: 40px;
+                }
+                .field-icon {
+                    position: absolute;
+                    right: 12px;
+                    top: 50%;
+                    transform: translate(0, -50%);
+                    width: 20px;
+                    cursor: pointer;
+                }
+            }
+            span.text-danger {
+                color: #dc3545 !important;
+                display: inline-block;
+                margin: 8px 0 0;
+                text-align: left;
+            }
+        }
+        .single-form-field.has-error {
+            input {
+                border-color: #dc3545 !important;
+            }
+        }
+        button.submit-button {
+            margin: 0;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            font-size: 1.25rem;
+            color: #fff;
+            background-color: #1885c9;
+            padding: 0.5rem 1rem;
+            border-radius: 5rem;
+            cursor: pointer;
+            width: 60%;
+            transition: all 200ms ease-in;
+            margin: 0 auto;
+        }
+    }
+}
+
+.v-middle {
+    position: absolute;
+    top: 50%;
+    transform: translate(0, -50%);
+}
+.v-h-middle {
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
+}
