+++ NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/components/common/SideMenuItems.js
@@ -1,0 +1,39 @@
+import React, { Component } from 'react'
+import { Link } from 'react-router-dom';
+import SlideToggle from 'react-slide-toggle';
+
+export default class SideMenuItems extends Component {
+    render() {
+        const menu = this.props.menu;
+
+        return (
+            <SlideToggle
+                collapsed
+                duration={300}
+                render={({ onToggle, setCollapsibleElement }) => (
+                    <>
+                        <div onClick={() => {
+                                onToggle();
+                            }}
+                            className="side-nav-menu link"
+                        >
+                            {menu.title}
+                        </div>
+
+                        <div 
+                            className="side-nav-submenu" 
+                            ref={setCollapsibleElement}
+                        >
+                            {
+                                menu.subNav.map((item, index) =>
+                                    <li key={index} onClick={this.props.hideSideNav}>
+                                        <Link className="side-nav-subnav link" to={item.path}>{item.title}</Link>
+                                    </li>)
+                            }
+                        </div>
+                    </>
+                )}
+            />
+        )
+    }
+}
