+++ NextGenSoftware.OASIS.API.ONODE.WebUI.HTML/react-app/src/index.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './Components/App';
-import './CSS/index.css';
+import App from './components/App';
+
+import 'bootstrap/dist/css/bootstrap.min.css';
+import 'font-awesome/css/font-awesome.min.css';
 
 ReactDOM.render(<App />, document.getElementById("root"));
